requirements.txt
Flask==1.1.2
keras==2.7.0
numpy==1.19.5
gunicorn==20.1.0
tensorflow-cpu==2.7.0
pillow==8.2.0